# api-docs.json
{"openapi":"3.1.0","info":{"title":"AudSys API - Sistema de gestión de auditorías","description":"API para la gestión de auditorías internas y externas de una empresa, observaciones, planes de acción y departamentos.\n\nAplicación de gestión de auditorías internas en Java con Spring Boot + React. Pensada para pequeñas empresas que quieran controlar sus procesos de calidad.","contact":{"name":"Alejandro Isabel","url":"https://www.alexisabel.com","email":"isabelsanchezalex@gmail.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"externalDocs":{"description":"Repositorio","url":"https://github.com/alexisabels/audsys"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Departamentos","description":"Gestión de departamentos auditables y sus responsables."},{"name":"Planes de Acción","description":"Planes definidos para corregir o mejorar aspectos detectados en las auditorías."},{"name":"Auditorías","description":"Operaciones relacionadas con auditorías"},{"name":"Auditores","description":"Gestión de auditores"},{"name":"Auditorías","description":"Operaciones relacionadas con las auditorías realizadas en los distintos departamentos."},{"name":"Observaciones","description":"Observaciones detectadas durante una auditoría."},{"name":"Auditores","description":"Gestión de los auditores que participan en las auditorías."},{"name":"Planes de acción","description":"Medidas tomadas ante hallazgos"},{"name":"Observaciones","description":"Anotaciones dentro de las auditorías"},{"name":"Departamentos","description":"Departamentos auditados"}],"paths":{"/api/v1/planAccion/{planId}":{"get":{"tags":["Planes de Acción"],"operationId":"getPlanAccion","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlanAccion"}}}}}},"put":{"tags":["Planes de Acción"],"operationId":"updatePlanAccion","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanAccionDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Planes de Acción"],"operationId":"deletePlan","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/observacion/{observacionId}":{"get":{"tags":["Observaciones"],"operationId":"getObservacion","parameters":[{"name":"observacionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Observacion"}}}}}},"put":{"tags":["Observaciones"],"operationId":"updateObservacion","parameters":[{"name":"observacionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ObservacionDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Observaciones"],"operationId":"deleteObservacion","parameters":[{"name":"observacionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/departamento/{departamentoId}":{"get":{"tags":["Departamentos"],"operationId":"getDepartamento","parameters":[{"name":"departamentoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Departamento"}}}}}},"put":{"tags":["Departamentos"],"operationId":"updateDepartamento","parameters":[{"name":"departamentoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartamentoDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Departamentos"],"operationId":"deleteDepartamento","parameters":[{"name":"departamentoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auditoria/{auditoriaId}":{"get":{"tags":["Auditorías"],"operationId":"getAuditoria","parameters":[{"name":"auditoriaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Auditoria"}}}}}},"put":{"tags":["Auditorías"],"operationId":"updateAuditoria","parameters":[{"name":"auditoriaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditoriaDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Auditorías"],"operationId":"deleteAuditoria","parameters":[{"name":"auditoriaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auditor/{auditorId}":{"get":{"tags":["Auditores"],"operationId":"getAuditor","parameters":[{"name":"auditorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Auditor"}}}}}},"put":{"tags":["Auditores"],"operationId":"updateAuditor","parameters":[{"name":"auditorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditorDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Auditores"],"operationId":"deleteAuditor","parameters":[{"name":"auditorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/planAccion":{"get":{"tags":["Planes de Acción"],"operationId":"getPlanes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlanAccion"}}}}}}},"post":{"tags":["Planes de Acción"],"operationId":"addPlanAccion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanAccionDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/observacion":{"get":{"tags":["Observaciones"],"operationId":"getObservaciones","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Observacion"}}}}}}},"post":{"tags":["Observaciones"],"operationId":"addObservacion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ObservacionDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/departamento":{"get":{"tags":["Departamentos"],"operationId":"getDepartamentos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Departamento"}}}}}}},"post":{"tags":["Departamentos"],"operationId":"addDepartamento","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartamentoDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auditoria":{"get":{"tags":["Auditorías"],"operationId":"getAuditorias","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Auditoria"}}}}}}},"post":{"tags":["Auditorías"],"operationId":"addAuditoria","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditoriaDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auditor":{"get":{"tags":["Auditores"],"operationId":"getAuditors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Auditor"}}}}}}},"post":{"tags":["Auditores"],"operationId":"addAuditor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditorDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"PlanAccionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descripcion":{"type":"string","minLength":1},"estado":{"type":"string","enum":["PENDIENTE","EN_CURSO","COMPLETADO","CANCELADO","RECHAZADO","APLAZADO","EN_REVISION","APROBADO","CERRADO"]},"auditoriaId":{"type":"integer","format":"int64"}},"required":["auditoriaId","descripcion","estado"]},"ObservacionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descripcion":{"type":"string"},"criticidad":{"type":"string","enum":["ALTA","MEDIA","BAJA"]},"auditoriaId":{"type":"integer","format":"int64"}}},"DepartamentoDTO":{"type":"object","properties":{"nombre":{"type":"string","minLength":1},"responsable":{"type":"string","minLength":1}},"required":["nombre","responsable"]},"AuditoriaDTO":{"type":"object","properties":{"fecha":{"type":"string","format":"date"},"tipo":{"type":"string","enum":["INTERNA","EXTERNA"]},"categorias":{"type":"array","items":{"type":"string","enum":["CALIDAD","SEGURIDAD","PROVEEDORES","REGULATORIA","INFORMAL"]},"minItems":1,"uniqueItems":true},"auditorId":{"type":"integer","format":"int64"},"departamentoId":{"type":"integer","format":"int64"},"observaciones":{"type":"array","items":{"$ref":"#/components/schemas/ObservacionDTO"}},"planesAccion":{"type":"array","items":{"$ref":"#/components/schemas/PlanAccionDTO"}}},"required":["auditorId","categorias","departamentoId","fecha","tipo"]},"AuditorDTO":{"type":"object","properties":{"nombre":{"type":"string"},"email":{"type":"string"}}},"Auditor":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"email":{"type":"string"}}},"Auditoria":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fecha":{"type":"string","format":"date"},"tipo":{"type":"string","enum":["INTERNA","EXTERNA"]},"categorias":{"type":"array","items":{"type":"string","enum":["CALIDAD","SEGURIDAD","PROVEEDORES","REGULATORIA","INFORMAL"]},"minItems":1,"uniqueItems":true},"departamento":{"$ref":"#/components/schemas/Departamento"},"auditor":{"$ref":"#/components/schemas/Auditor"},"observaciones":{"type":"array","items":{"$ref":"#/components/schemas/Observacion"}},"planesAccion":{"type":"array","items":{"$ref":"#/components/schemas/PlanAccion"}}},"required":["auditor","categorias","departamento","fecha","tipo"]},"Departamento":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"responsable":{"type":"string"}}},"Observacion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descripcion":{"type":"string","minLength":1},"criticidad":{"type":"string","enum":["ALTA","MEDIA","BAJA"]},"auditoria":{"$ref":"#/components/schemas/Auditoria"}},"required":["auditoria","criticidad","descripcion"]},"PlanAccion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"descripcion":{"type":"string"},"estado":{"type":"string","enum":["PENDIENTE","EN_CURSO","COMPLETADO","CANCELADO","RECHAZADO","APLAZADO","EN_REVISION","APROBADO","CERRADO"]},"auditoria":{"$ref":"#/components/schemas/Auditoria"}}}}}}